requirements.txt
Django==5.2.5
djangorestframework
django-environ
django-cors-headers
psycopg2-binary
gunicorn
celery
redis
pytest
pytest-django
pre-commit
